+++ src/systems/pf2e.js
@@ -282,7 +282,6 @@ export class pf2e{
     }
 
     async toggleCondition(token,condition) {
-        console.log(token,condition)
         if (condition == undefined) condition = 'removeAll';
         if (condition == 'removeAll'){
             for( let effect of token.actor.items.filter(i => i.type == 'condition'))
@@ -290,7 +289,6 @@ export class pf2e{
         }
         else {
             const effect = this.getCondition(token,condition);
-            console.log('eff',effect)
             if (effect == undefined) {
                 const newEffect = game.pf2e.ConditionManager.conditions.get(condition).toObject();
                 await token.actor?.createEmbeddedDocuments("Item", [newEffect]);